+++ src/quanlythuchi/services/ConnDatabase.java
@@ -107,7 +107,7 @@ public class ConnDatabase {
             Statement st = conn.createStatement();
             ResultSet rs = st.executeQuery(sql);
             while (rs.next()) {
-                list.add(new GiaoDich(rs.getString(1), rs.getInt(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getInt(6), rs.getInt(7)));
+                list.add(new GiaoDich(rs.getString(1), rs.getDouble(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getInt(6), rs.getInt(7)));
             }
             rs.close();
             st.close();
@@ -124,7 +124,7 @@ public class ConnDatabase {
             Statement st = conn.createStatement();
             ResultSet rs = st.executeQuery(sql);
             while (rs.next()) {
-                list.add(new GiaoDich(rs.getString(1), rs.getInt(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getInt(6), rs.getInt(7)));
+                list.add(new GiaoDich(rs.getString(1), rs.getDouble(2), rs.getString(3), rs.getString(4), rs.getString(5), rs.getInt(6), rs.getInt(7)));
             }
             rs.close();
             st.close();
